Abstract common functionality into a Base class
The base class understands how to generate an arbitarily complex WHERE clause.
It can handle constructs such as Lists of other Table objects, and so forth.

Additionally, the types allowable in the WHERE filter are now specified in the
type system and are available as a simple import from the `orm.abc` module.